requirements.txt
# WebSocket
websockets>=12.0

# HTTP client for skills
aiohttp>=3.9.0

# ASR - OpenAI Whisper
openai-whisper>=20231117

# LLM
openai>=1.0.0

# Audio processing
pyaudio>=0.2.14
numpy>=1.24.0

# Config
python-dotenv>=1.0.0

# TTS - CosyVoice 依赖
# 需要先克隆 CosyVoice: git clone --recursive https://github.com/FunAudioLLM/CosyVoice.git
# 然后 cd CosyVoice && pip install -r requirements.txt
funasr>=1.0.0
onnxruntime>=1.16.0
webrtcvad>=2.0.10
modelscope>=1.10.0
